requirements.txt
aiohttp>=3.9.0
dnspython>=2.4.2
